[SQL] Make sure retain operators are applied to the correct integrators
