+++ sql-to-dbsp-compiler/SQL-compiler/src/main/java/org/dbsp/sqlCompiler/circuit/operator/DBSPIntegrateTraceRetainValuesOperator.java
@@ -78,4 +78,6 @@ public final class DBSPIntegrateTraceRetainValuesOperator
             visitor.postorder(this);
         visitor.pop(this);
     }
+
+    // equivalent inherited from parent
 }
